+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<!-- next-header -->
 
-## [Unreleased]
+## [Unreleased] - ReleaseDate
 ### Added
 - compression: implement Lz4/Lz4Hc compression modes for `INSERT`s ([#39]).
 - insert: feature `wa-37420` to avoid [ClickHouse#37420].
@@ -192,7 +193,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Client::insert()` for inserting into tables.
 - `Client::query()` for selecting from tables and DDL statements.
 
-[unreleased]: https://github.com/loyd/clickhouse.rs/compare/v0.10.0...HEAD
+<!-- next-url -->
+[Unreleased]: https://github.com/loyd/clickhouse.rs/compare/v0.10.0...HEAD
 [0.10.0]: https://github.com/loyd/clickhouse.rs/compare/v0.9.3...v0.10.0
 [0.9.3]: https://github.com/loyd/clickhouse.rs/compare/v0.9.2...v0.9.3
 [0.9.2]: https://github.com/loyd/clickhouse.rs/compare/v0.9.1...v0.9.2